// src/types/APIAmiInterfaces.ts
export interface VisitasHospitales {
    Id: string;
    createdAt?: Date;
    updatedAt?: Date;
    deleted?: string;
    IdHospital?: string;
    DscHospital?: string;
    idPaciente?: string;
    NombrePaciente: string;
    IdTerapeuta?: string;
    NombreTerapeuta?: string;
    IdMesVisita?: string;
    DscMesVisita?: string;
    FechaVisita?: string ;
    IdTipoVisita?: string;
    DscTipoVisita?: string;
    IdTipoEquipo?: string;
    DscTipoEquipo?: string;
    FrecuenciaCardiaca?: string;
    PresionArterialSistolica?: string;
    PresionArterialDiastolica?: string;
    SaturacionOxigeno?: string;
    HoraUsoPromDia?: string;
    HorasTotalMensuales?: string;
    DiasUsoSobreTotal?: string;
    FugaLmin?: string;
    IndiceApnea?: string;
    PresionUtilizadaEpap?: string;
    PresionUtilizadaIPAP?: string;
    PresionUtilizadaCPAP?: string;
    CambioEquipo?: boolean;
    NumSerieEquipoyDN?: string;
    IdEstadoEquipo?: string;
    DscEstadoEquipo?: string;
    IdEvaluacionVisita?: string;
    DscEvaluacionVisitas?: string;
    ObservacionesClinicas?: string;
    ComentariosAdministrativos?: string;
    SharePoint?: boolean;
    EstadoPaciente : string,
    DsEstadoPaciente : string,
    AdherenciaUsoTratamiento: string,
    DsAdherenciaUsoTratamiento: string,
    CambEquiInsuTrata: string,
    DsCambEquiInsuTrata: string,
    CondTrataYParam: string,
    DsCondTrataYParam: string,
    SituaEspecYCoordi: string,
    DsSituaEspecYCoordi: string,

  }